SONARPY-2284 Add tests for S2638 on imported class with decorator
